requirements.txt
requests
jmespath
box-sdk-gen
azure-cosmos
csv